requirements_pytorch.txt
torch>=2.0.0
torchvision>=0.15.0
grad-cam>=1.4.8
captum>=0.6.0
shap>=0.44.0
lime>=0.2.0.1
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
pandas>=2.0.0
tqdm>=4.65.0
kaggle>=1.5.13
